requirements.txt
pyparsing==2.1.5
python-dateutil==2.5.3
psutil==4.3.0
multiprocess==0.70.5

# posix
pybloomfiltermmap==0.3.15
# windows
pybloom==1.1